Move pyspellchecker interactive installation flow from configpanel to SpellCheckManager
